bot/pyproject.toml
# uv workspace for the bot

[project]
name = "restaurant-bot"
version = "0.1.0"
description = "Restaurant Allergen Telegram Bot"
requires-python = ">=3.12"
dependencies = [
  "python-telegram-bot==21.7",
  "httpx==0.28",
  "pydantic-settings==2.6",
  "sqlmodel==0.0.22",
]

[dependency-groups]
dev = ["pytest==8.3", "pytest-asyncio==0.24"]
